requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
fastapi
uvicorn
qdrant-client
sentence-transformers
pydantic
llama-cpp-python
